+++ src/controllers/contacts/CreateContacts.js
@@ -1,15 +1,7 @@
 import moment from 'moment';
 import contacts from '../../model/contacts';
-import { validateContact } from '../../helpers/validations/contact';
 
 export const createContacts = (req, res) => {
-  const { error } = validateContact(req.body);
-  if (error) {
-    return res.status(400).send({
-      status: 400,
-      error: error.details[0].message
-    });
-  }
   const contact = {
     id: parseInt(contacts.length + 1, 10),
     createdOn: moment().format('LL'),
