+++ src/controllers/messages/deleteMessage.js
@@ -1,6 +1,5 @@
 import messages from '../../model/messages';
 
-
 export const deleteMessage = (req, res) => {
   const messageIndex = messages.findIndex(message => message.id === parseInt(req.params.id, 10));
   if (messageIndex > -1) {
