+++ src/controllers/messages/getMessages.js
@@ -23,7 +23,7 @@ export const getUnreadMessages = (req, res) => {
 
 export const getReadMessages = (req, res) => {
   const read = messages.filter(message => message.status === 'read');
-  if (read) {
+  if (read.length > 0) {
     return res.status(200).send({
       status: 200,
       data: [read]
@@ -31,6 +31,22 @@ export const getReadMessages = (req, res) => {
   }
   return res.status(404).send({
     status: 404,
-    message: 'The inbox message was not found'
+    message: 'No message found'
+  });
+};
+
+
+export const getSpecificMessagesForUser = (req, res) => {
+  const select = messages.find(message => message.receiverId === parseInt(req.params.contactId, 10));
+  if (select.length > -1) {
+    return res.status(200).send({
+      status: 200,
+      data: [select]
+    });
+  }
+
+  return res.status(404).send({
+    status: 404,
+    error: `The message for the contact with id ${req.params.contactId} is not found`
   });
 };
