+++ src/helpers/validations/contact.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
 
 const Contact = {
-  async validate(req, res) {
+  async validate(req, res, next) {
     const schema = Joi.object().keys({
       email: Joi.string().email({ minDomainAtoms: 2 }),
       firstname: Joi.string().alphanum().min(3).max(30)
@@ -17,7 +17,7 @@ const Contact = {
         error: error.details[0].message
       });
     }
-    return res.json(value);
+    next();
   },
 };
